refactor: upload 도메인의 상수값과 타입을 따로 분리
상수값과 타입을 따로 분리

#171
