+++ FE/src/Model/Upload.ts
@@ -1,0 +1,6 @@
+export type UploadParams = "info" | "option" | "plan" | "content" | "creator";
+
+export type UploadNavList = {
+  title: string;
+  path: UploadParams;
+};
