+++ FE/src/components/pages/new/templates/UploadPageNavbar.tsx
@@ -1,33 +1,6 @@
+import { uploadNavLists } from "@/constants/uploadData";
 import { cn } from "@/utils/cn";
 import Link from "next/link";
-
-type UploadNavList = {
-  title: string;
-  path: "info" | "option" | "plan" | "content" | "creator";
-};
-
-const uploadNavLists: UploadNavList[] = [
-  {
-    title: "펀딩 정보",
-    path: "info",
-  },
-  {
-    title: "옵션 설정",
-    path: "option",
-  },
-  {
-    title: "펀딩 계획",
-    path: "plan",
-  },
-  {
-    title: "내용 작성",
-    path: "content",
-  },
-  {
-    title: "크리에이터 정보",
-    path: "creator",
-  },
-];
 
 type Props = {
   currentPath: "info" | "option" | "plan" | "content" | "creator";
