+++ FE/src/constants/uploadData.ts
@@ -1,0 +1,24 @@
+import { UploadNavList } from "@/Model/Upload";
+
+export const uploadNavLists: UploadNavList[] = [
+  {
+    title: "펀딩 정보",
+    path: "info",
+  },
+  {
+    title: "옵션 설정",
+    path: "option",
+  },
+  {
+    title: "펀딩 계획",
+    path: "plan",
+  },
+  {
+    title: "내용 작성",
+    path: "content",
+  },
+  {
+    title: "크리에이터 정보",
+    path: "creator",
+  },
+];
